Placement phase refactor
