+++ Assets/Scripts/PlacementPhase.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 651c22ed39b3cad0886869557d6dc555
+guid: cc7bc440e899a5e1aa7473ac7adf2244
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
